se reorganizan todas los archivos de entities
