+++ src/entities/category.entity.ts
@@ -1,6 +1,6 @@
 import { Column, Entity, OneToMany } from "typeorm";
 import { BaseEntity } from "../config/base.entity";
-import { ProductEntity } from "../product/product.entity";
+import { ProductEntity } from "./product.entity";
 
 @Entity({ name: "category" })
 export class CategoryEntity extends BaseEntity {
